backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
motor
pydantic-settings
passlib
python-jose
email-validator
python-multipart
fastapi==0.111.0
uvicorn[standard]==0.29.0

motor==3.4.0
pymongo==4.15.5

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

google-auth==2.29.0
google-auth-oauthlib==1.2.0

cloudinary==1.44.1
python-multipart==0.0.9

pydantic==2.10.6
pydantic-settings==2.6.1

email-validator==2.1.1
httpx==0.27.0
python-dotenv==1.0.1
python-dateutil==2.9.0
